prepare single-camera driver for use with sync'ed driver
